requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Talisman==1.1.0
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7
WTForms==3.0.1
gunicorn==21.2.0
requests==2.31.0
google-cloud-speech==2.21.0
google-generativeai==0.3.2
openai==1.3.0
reportlab==4.0.4
Pillow==10.0.1
google-cloud-translate==3.11.1
google-cloud-texttospeech==2.16.3
